requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.5
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-celery-results==2.0.1
django-redis==5.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-websocket-redis==0.6.0
djangorestframework==3.12.4
dnspython==1.16.0
eventlet==0.31.0
gevent==21.1.2
greenlet==1.1.0
idna==2.10
importlib-metadata==4.5.0
kombu==5.1.0
oauthlib==3.1.1
pkg-resources==0.0.0
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.16.0
sqlparse==0.4.1
stripe==2.58.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
